cl registerServices restructure
