+++ lib/firebase/register/ui_helper.dart
@@ -1,0 +1,26 @@
+import 'package:flutter/material.dart';
+
+class UIHelper {
+  static void showSuccessMessage(BuildContext context,
+      {String message = 'Registration successful'}) {
+    if (!context.mounted) return;
+    ScaffoldMessenger.of(context).showSnackBar(
+      SnackBar(
+        content: Text(message),
+        backgroundColor: Colors.green,
+        behavior: SnackBarBehavior.floating,
+      ),
+    );
+  }
+
+  static void showErrorMessage(BuildContext context, String message) {
+    if (!context.mounted) return;
+    ScaffoldMessenger.of(context).showSnackBar(
+      SnackBar(
+        content: Text(message),
+        backgroundColor: Colors.red,
+        behavior: SnackBarBehavior.floating,
+      ),
+    );
+  }
+}
