+++ lib/pages/loading/login_register/temp_login_page.dart
@@ -267,7 +267,7 @@ class _LoginPageState extends State<LoginPage> {
     Navigator.pushReplacement(
       context,
       PageRouteBuilder(
-        pageBuilder: (context, animation, _) => const TempRegisterPage(),
+        pageBuilder: (context, animation, _) => const RegisterPage(),
         transitionsBuilder: (context, animation, secondaryAnimation, child) {
           return SlideTransition(
             position: animation.drive(
